+++ src/main/java/tikape/runko/database/RaakaaineDao.java
@@ -66,6 +66,11 @@ public class RaakaaineDao implements Dao<Raakaaine, Integer> {
 
         return opiskelijat;
     }
+    
+    @Override
+    public Raakaaine saveOrUpdate(Raakaaine object) throws SQLException {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
 
     @Override
     public void delete(Integer key) throws SQLException {
