+++ src/main/java/tikape/runko/domain/Annos.java
@@ -1,4 +1,4 @@
-package tikape.runko.database;
+package tikape.runko.domain;
 
 public class Annos {
